pyproject.toml
[project]
name = "wpa_recommendations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.11"
dependencies = [
    "aiohttp-client-cache[all]>=0.12.4",
    "aiohttp>=3.11.8",
    "asyncpg>=0.30.0",
    "backoff>=2.2.1",
    "disnake-jishaku==2.6.5",
    "disnake==2.9.2",
    "ipykernel>=6.29.5",
    "pybind11[global]>=2.13.6",
    "python-dotenv>=1.0.1",
    "qdrant-client>=1.12.1",
    "rich>=13.9.4",
]

[tool.ruff.lint]
ignore = ['E402']
